requirements.txt
h5py
matplotlib
numpy
Pillow
Requests
streamlit
streamlit_lottie
tensorflow
tensorflow_hub
tensorflow_intel